* Refactoring: move sink/source buffering into separate classes. * Buffer the HashSink.  This speeds up hashing a bit because it   prevents lots of calls to the hash update functions (e.g. nix-hash   went from 9.3s to 8.7s of user time on the closure of my   /var/run/current-system).

+++ src/libstore/references.cc
@@ -57,11 +57,11 @@ struct RefScanSink : Sink
 
     RefScanSink() : hashSink(htSHA256) { }
     
-    void operator () (const unsigned char * data, unsigned int len);
+    void operator () (const unsigned char * data, size_t len);
 };
 
 
-void RefScanSink::operator () (const unsigned char * data, unsigned int len)
+void RefScanSink::operator () (const unsigned char * data, size_t len)
 {
     hashSink(data, len);
